Initial Upload
CRC correctly generates. Message and CRC size are manually set before simulation in source and testbench files. Commented sections under "use for an accurate schematic" are included. When uncommented (and the "use for brevity" sections are commented), the generated schematic is a little more accurate, showing multiple D Flip-Flops instead of a D Flip-Flop register. The number of CRC bit signals must be manually adjusted then based on the desired CRC size.